링크 카드 디자인 수정
