+++ src/component/App/index.js
@@ -1,6 +1,6 @@
 import "./App.css";
-import Navigation from "../Navigation/Navigation";
-import Footer from "../Footer/Footer";
+import Navigation from "../Navigation";
+import Footer from "../Footer";
 import { Outlet } from "react-router-dom";
 
 function App() {
