--- src/component/LinkItems/InputContiner.js
@@ -1,7 +1,0 @@
-import styled from "styled-components";
-
-const InputContainer = styled.div`
-  grid-column: 1 / -1;
-`;
-
-export default InputContainer;
