+++ src/component/LinkItems/LinkItems.css
@@ -17,7 +17,7 @@
 .LinkItem {
   width: 340px;
   height: 334px;
-  border-radius: 8px;
+  border-radius: 1rem;
   box-shadow: 0px 5px 25px 0px rgba(0, 0, 0, 0.08);
 }
